refactor(notifications-backfill): split into types/api/github/loop/process modules
Brings packages/crane-mcp/src/scripts/notifications-backfill.ts under
structural caps (was 718 LOC).

- notifications-backfill.ts (259 LOC): runBackfill delegates to helpers,
  within 75-line and complexity-15 ceilings.
- notifications-backfill/types.ts: interface/type definitions.
- notifications-backfill/crane-api.ts: lock acquire/release, pending-matches
  query, open-notifications query, auto-resolve POST.
- notifications-backfill/github.ts: GitHub Actions API fetch with rate-limit
  handling and Link-header pagination.
- notifications-backfill/loop.ts: runMatchesLoop (while-true page walker) +
  processPageMatches.
- notifications-backfill/process-match.ts: processMatch + resolveNotification
  (auto-resolve block separated from dry-run branch).

Public API preserved. 13/13 tests pass.
